Enhanced the basic window and renderer structure
-Added GLRenderWindow which inherits from RenderWindow which in turn from RenderTarget.
GLRenderWindow is the one instatiated in the scene. It takes its own buffer (provided by ancestor RenderTarget)
and splats it onto the screen as a texture

-Added hierarchy of renderers. Root of the hierarchy is BufferedRenderer, which is an interface for rendering
into arbitrary memory buffer. RaytracingRenderer inherits from the BufferedRenderer and gives it general ray tracing
capabilities (currently only init). From RaytracingRenderer several instantiable renderers are created:
cudaVolumeRenderer and rtfactRenderer so far. This should enable switching renderers in run-time

+++ include/BALL/VIEW/RENDERING/renderWindow.h
@@ -96,7 +96,7 @@ namespace BALL
             const unsigned int m_minimalHeight;     // minimum height the window can have
 		};
 
-        typedef RenderWindow<int> t_RenderWindow;
+        typedef RenderWindow<float> t_RenderWindow;
 
 	} // namespace VIEW
